refacto advanced search
